Activate PC and register file on falling edge
Simulated memory reads are asynchronous, so moving writes to the falling
edge emulates writing on the first half of the clock cycle and reading
on the second half.

This change is only relevant for PC and the register file due to branch
and forward, which both modify values which should be written to the
registers before the next rising edge.