Fix branch address resolution
Requires propagating the instruction immediate and PC up to the MEM
state, in which PC is updated with its old value added to the shifted
immediate.

Propagating PC is required since, due to branch prediction, no stall
occurs after a branch, so PC has already been updated twice by the time
the branch address is calculated.