Fix control hazard flushing
Flushes EX/MEM registers rather than IF/ID ones.

Due to writing to PC on the falling edge, the correct instruction has
already been fetched when IF/ID registers it, so there is no need to
flush it. However, it is necessary to flush EX/MEM since the incorrect
instruction is propagated to it at the same time.